extract MethodCallTransformer into separate file + minor refactoring
