+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/transformation/LincheckClassVisitor.kt
@@ -950,227 +950,6 @@ internal class LincheckClassVisitor(
     }
 
     /**
-     * Adds strategy method invocations before and after method calls.
-     */
-    private inner class MethodCallTransformer(
-        fileName: String,
-        className: String,
-        methodName: String,
-        adapter: GeneratorAdapter,
-    ) : ManagedStrategyMethodVisitor(fileName, className, methodName, adapter) {
-        override fun visitMethodInsn(opcode: Int, owner: String, name: String, desc: String, itf: Boolean) = adapter.run {
-                // TODO: do not ignore <init>
-                if (isCoroutineInternalClass(owner)) {
-                    invokeInIgnoredSection {
-                        visitMethodInsn(opcode, owner, name, desc, itf)
-                    }
-                    return
-                }
-                if (owner == "sun/misc/Unsafe" ||
-                    owner == "jdk/internal/misc/Unsafe" ||
-                    owner == "java/lang/invoke/VarHandle" ||
-                    owner.startsWith("java/util/concurrent/") && (owner.contains("Atomic")) ||
-                    owner.startsWith("kotlinx/atomicfu/") && (owner.contains("Atomic"))
-                ) {
-                    invokeIfInTestingCode(
-                        original = {
-                            visitMethodInsn(opcode, owner, name, desc, itf)
-                        },
-                        code = {
-                            processAtomicMethodCall(desc, opcode, owner, name, itf)
-                        }
-                    )
-                    return
-                }
-                if (
-                    name == "<init>" ||
-                    owner.startsWith("sun/nio/ch/lincheck/") ||
-                    owner.startsWith("org/jetbrains/kotlinx/lincheck/") ||
-                    owner == "kotlin/jvm/internal/Intrinsics" ||
-                    owner == "java/util/Objects" ||
-                    owner == "java/lang/String" ||
-                    owner == "java/lang/Boolean" ||
-                    owner == "java/lang/Long" ||
-                    owner == "java/lang/Integer" ||
-                    owner == "java/lang/Short" ||
-                    owner == "java/lang/Byte" ||
-                    owner == "java/lang/Double" ||
-                    owner == "java/lang/Float" ||
-                    owner == "java/util/Locale" ||
-                    owner == "org/slf4j/helpers/Util" ||
-                    owner == "java/util/Properties"
-                ) {
-                    visitMethodInsn(opcode, owner, name, desc, itf)
-                    return
-                }
-                invokeIfInTestingCode(
-                    original = {
-                        visitMethodInsn(opcode, owner, name, desc, itf)
-                    },
-                    code = {
-                        val endLabel = newLabel()
-                        val methodCallStartLabel = newLabel()
-                        // STACK [INVOKEVIRTUAL]: owner, arguments
-                        // STACK [INVOKESTATIC]: arguments
-                        val argumentLocals = storeArguments(desc)
-                        // STACK [INVOKEVIRTUAL]: owner
-                        // STACK [INVOKESTATIC]: <empty>
-                        when (opcode) {
-                            INVOKESTATIC -> visitInsn(ACONST_NULL)
-                            else -> dup()
-                        }
-                        push(owner)
-                        push(name)
-                        loadNewCodeLocationId()
-                        // STACK [INVOKEVIRTUAL]: owner, owner, className, methodName, codeLocation
-                        // STACK [INVOKESTATIC]:         null, className, methodName, codeLocation
-                        val argumentTypes = getArgumentTypes(desc)
-                        push(argumentLocals.size) // size of the array
-                        visitTypeInsn(ANEWARRAY, OBJECT_TYPE.internalName)
-                        // STACK: ..., array
-                        for (i in argumentLocals.indices) {
-                            // STACK: ..., array
-                            dup()
-                            // STACK: ..., array, array
-                            push(i)
-                            // STACK: ..., array, array, index
-                            loadLocal(argumentLocals[i])
-                            // STACK: ..., array, array, index, argument[index]
-                            box(argumentTypes[i])
-                            arrayStore(OBJECT_TYPE)
-                            // STACK: ..., array
-                        }
-                        // STACK: ..., array
-                        invokeStatic(Injections::beforeMethodCall)
-                        invokeBeforeEventIfPluginEnabled("method call $methodName", setMethodEventId = true)
-                        // STACK [INVOKEVIRTUAL]: owner, arguments
-                        // STACK [INVOKESTATIC]: arguments
-                        val methodCallEndLabel = newLabel()
-                        val handlerExceptionStartLabel = newLabel()
-                        val handlerExceptionEndLabel = newLabel()
-                        visitTryCatchBlock(methodCallStartLabel, methodCallEndLabel, handlerExceptionStartLabel, null)
-                        visitLabel(methodCallStartLabel)
-                        loadLocals(argumentLocals)
-                        visitMethodInsn(opcode, owner, name, desc, itf)
-                        visitLabel(methodCallEndLabel)
-                        // STACK [INVOKEVIRTUAL]: owner, arguments
-                        // STACK [INVOKESTATIC]: arguments
-                        val resultType = getReturnType(desc)
-                        if (resultType == VOID_TYPE) {
-                            invokeStatic(Injections::onMethodCallVoidFinishedSuccessfully)
-                        } else {
-                            val resultLocal = newLocal(resultType)
-                            storeLocal(resultLocal)
-                            loadLocal(resultLocal)
-                            box(resultType)
-                            invokeStatic(Injections::onMethodCallFinishedSuccessfully)
-                            loadLocal(resultLocal)
-                        }
-                        // STACK: value
-                        goTo(handlerExceptionEndLabel)
-                        visitLabel(handlerExceptionStartLabel)
-                        dup()
-                        invokeStatic(Injections::onMethodCallThrewException)
-                        throwException()
-                        visitLabel(handlerExceptionEndLabel)
-                        // STACK: value
-                        visitLabel(endLabel)
-                    }
-                )
-            }
-
-        private fun GeneratorAdapter.processAtomicMethodCall(
-            desc: String,
-            opcode: Int,
-            owner: String,
-            name: String,
-            itf: Boolean,
-        ) {
-            // In cases of Atomic*FieldUpdater, Unsafe and VarHandle we edit
-            // the params list before creating a trace point to remove redundant parameters
-            // as receiver and offset.
-            // To determine how we should process it, we provide owner instance.
-            val isVarHandleOrAtomicReference = (owner == "java/lang/invoke/VarHandle" ||
-                    owner == "java/util/concurrent/atomic/AtomicLong" ||
-                    owner == "java/util/concurrent/atomic/AtomicInteger" ||
-                    owner == "java/util/concurrent/atomic/AtomicBoolean" ||
-                    (owner == "java/util/concurrent/atomic/AtomicReference") ||
-                    owner == "java/util/concurrent/atomic/AtomicReferenceArray" ||
-                    owner == "java/util/concurrent/atomic/AtomicIntegerArray" ||
-                    owner == "java/util/concurrent/atomic/AtomicLongArray") && name != "<init>"
-            val provideOwner = opcode != INVOKESTATIC &&
-                    (owner.endsWith("FieldUpdater") ||
-                            owner == "sun/misc/Unsafe" ||
-                            owner == "jdk/internal/misc/Unsafe" ||
-                            isVarHandleOrAtomicReference)
-
-            // STACK [INVOKEVIRTUAL]: owner, arguments
-            // STACK [INVOKESTATIC]: arguments
-            val argumentLocals = storeArguments(desc)
-            // STACK [INVOKEVIRTUAL]: owner
-            // STACK [INVOKESTATIC]: <empty>
-            if (provideOwner) {
-                dup()
-            } else {
-                visitInsn(ACONST_NULL)
-            }
-            // STACK [INVOKEVIRTUAL]: owner, owner
-            // STACK [INVOKESTATIC]:       , null
-            push(owner)
-            // STACK [INVOKEVIRTUAL]: owner, owner, className
-            // STACK [INVOKESTATIC]:         null, className,
-            push(name)
-            loadNewCodeLocationId()
-            // STACK [INVOKEVIRTUAL]: owner, owner, className, methodName, codeLocation
-            // STACK [INVOKESTATIC]:         null, className, methodName, codeLocation
-
-            // STACK [INVOKEVIRTUAL]: owner, ownerName, className, methodName, codeLocation
-            // STACK [INVOKESTATIC]:         ownerName, className, methodName, codeLocation
-            val argumentTypes = getArgumentTypes(desc)
-
-            push(argumentLocals.size) // size of the array
-            visitTypeInsn(ANEWARRAY, OBJECT_TYPE.internalName)
-            // STACK: ..., array
-            for (i in argumentLocals.indices) {
-                // STACK: ..., array
-                dup()
-                // STACK: ..., array, array
-                push(i)
-                // STACK: ..., array, array, index
-                loadLocal(argumentLocals[i])
-                // STACK: ..., array, array, index, argument[index]
-                box(argumentTypes[i])
-                arrayStore(OBJECT_TYPE)
-                // STACK: ..., array
-            }
-            // STACK: ..., array
-            invokeStatic(Injections::beforeAtomicMethodCall)
-            invokeBeforeEventIfPluginEnabled("atomic method call $methodName")
-
-            // STACK [INVOKEVIRTUAL]: owner, arguments
-            // STACK [INVOKESTATIC]: arguments
-            loadLocals(argumentLocals)
-            invokeInIgnoredSection {
-                visitMethodInsn(opcode, owner, name, desc, itf)
-            }
-            // STACK [INVOKEVIRTUAL]: owner, arguments
-            // STACK [INVOKESTATIC]: arguments
-            val resultType = getReturnType(desc)
-            if (resultType == VOID_TYPE) {
-                invokeStatic(Injections::onMethodCallVoidFinishedSuccessfully)
-            } else {
-                val resultLocal = newLocal(resultType)
-                storeLocal(resultLocal)
-                loadLocal(resultLocal)
-                box(resultType)
-                invokeStatic(Injections::onMethodCallFinishedSuccessfully)
-                loadLocal(resultLocal)
-            }
-            // STACK: value
-        }
-    }
-
-    /**
      * Adds invocations of ManagedStrategy methods before wait and after notify calls
      */
     private inner class WaitNotifyTransformer(
